This example is buggy now.
Use page `posts/del/:id` as example.
If you go to the page directly (copy and paste the url to the browser),
the server side-syncer would be triggered.
On the other hand, if you click through the page, then the client-side syncer would be triggered.
The bugs are aucsed by the difference between the syncers:
1. The `:id` param could be got by `req.param('id')` when triggering server-side syncer.
   But when triggering client-side syncer, `req.param('id')` returns nothing.
2. If triggering client-side syncer, then calling `getTemplateDate` of the view would return the result from the syncer.
   But it returns undefined when using server-side syncer.